Conflicts:
	twittshell.py

+++ twittshell.py
@@ -31,15 +31,19 @@
 #  OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF 
 #  ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#sorry I'm a C guy commands shall end with ; as long as I'm here 
-# that's pretty stupid, you should be sorry about it
+#Python: sorry I'm a C guy commands shall end with ; as long as I'm here 
+#  BatSteve: that's pretty stupid, you should be sorry about it
+#Python: your code dosn't have too, however as long as I have to use C it wreaks havoc 
+# on my code if I don't maintain such programming practices. Anyways look on the 
+# bright side python doesn't even see the semicolons's or unnecessary parentheses
+# Though I should use true and false more often in my code.
 
 import time
 import twitter
 import sys
 import getopt
 
-
+#default: python50 my test account
 consumer_key = '2x0h71VcEAk2dAdf886HmA';
 consumer_secret = 'WWsaaEJOzCuRtUxhR2kU6LShZNyNBSufuKod91N1U';
 access_key = '254290857-CrESL7rXpjMlJnaCqlTwANYS2J4dRObMbdqLTcM';
@@ -56,29 +60,20 @@ api = twitter.Api(consumer_key=consumer_key, consumer_secret=consumer_secret,
 
 message='';
 last_deleted_message='';
-last_status = twitter.Status();
+last_status = None; #twitter.Status();
 current_timezone=(time.timezone/60)/60; # Timezone
-status_update_delay=0;#api.MaximumHitFrequency()
+status_update_delay=0;# 0 or api.MaximumHitFrequency() <- 8
 last_reply_id=0;
 
-version_number = "0.1c RC1";
+version_number = "0.14";
 
+#------------------------------------------------------------
 #printl appends carriage returns ...
 def printl(string):
 	print (string+"\r");
 
 #------------------------------------------------------------
-#time date functions
-def current_time_date():
-	timen=time.gmtime();
-	return str(timen[3]-current_timezone)+":"+str(timen[4])+" "+str(timen[1])+"-"+str(timen[2])+"-"+str(timen[0]);
-
-def current_time():
-	timen=time.gmtime();
-	return str(timen[3]-current_timezone)+":"+str(timen[4]);
-
-#------------------------------------------------------------
-#usage
+#usage text
 def usage():
 	printl("");
 	printl("Usage: twitshell.py [-p message_content] [-r]");
@@ -89,16 +84,17 @@ def usage():
 	printl("	-r		Read(and Display) recent replies --read");
 
 #------------------------------------------------------------
-#About
+#about text
 def about():
 	printl("Twitter Shell Version "+version_number);
 	printl("	By Jason White, February 20 2011");
 	printl("");
 	printl("Special Thanks to:");
 	printl("	Alphageek");
-	printl("	Quick_Ben");
+	printl("	BatSteve");# why don't you ever add yourself ?
 	printl("	maglinvinn");
 	printl("	MrTransistor");
+	printl("	Quick_Ben");
 	printl("	Sparky Projects");
 	printl("	Tmb");
 	printl("	Tysk");
@@ -112,59 +108,72 @@ def about():
 	printl("This Program is licensed under the Modified BSD license");
 	printl("	Program Copyright (C) 2011 Jason White\n");
 #------------------------------------------------------------
-#The Banner
+#banner text
 def banner():
 	printl("	Twitter Shell Version "+version_number);
 	printl(" Because the geek shall inherit the earth");
 	printl("");
 
 #------------------------------------------------------------
-#The Bad command text, this could be made into a var but its cleaner as a function
-def bad_cmd():
-	#banner()
+#bad command text
+def bad_command():
 	printl("Unknown command type ? or help for help");
 
 #------------------------------------------------------------
 #Help Me ! Text
 def help():
-	printl("Help Me ! - Commands");
+	printl("Help Me ! - Commands:");
 	printl("");
-	printl("	about - more info about this program");
-	printl("	delete - delete the last message");
-	printl("	help - printl(this help text");
-	printl("	read - read replies");
-	printl("	undelete - undelete the last post, if it was deleted");
-	printl("	post - post a message to twitter");
-	printl("	quit - quit this program");
+	printl("	[a] or about	- more info about this program");
+	printl("	[d] or delete	- delete the last message");
+	printl("	[h] or help	- this help text");
+	printl("	[r] or read	- read replies");
+	printl("	[u] or undelete	- undelete the last post, if it was deleted");
+	printl("	[p] or post	- post a message to twitter");
+	printl("	[q] or quit	- quit this program");
 	printl("");
-	printl("	Twitter Shell"+version_number+"For The Geek Group");
+	printl("	Twitter Shell "+version_number+" For The Geek Group");
 	printl("	 By Jason White February 20 2011\n");
 
 #------------------------------------------------------------
-# Gets input from the commandline
+#time date functions, returns string
+def current_time_date():
+	timen=time.gmtime();
+	return str(timen[3]-current_timezone)+":"+str(timen[4])+" "+str(timen[1])+"-"+str(timen[2])+"-"+str(timen[0]);
+
+def current_time():
+	timen=time.gmtime();
+	return str(timen[3]-current_timezone)+":"+str(timen[4]);
+
+#------------------------------------------------------------
+# command: Gets input from the command line
 def command():
 	cmd=str(raw_input(current_time()+" Command >")).lower();  # make the command all lowercase for usage
 	time.sleep(status_update_delay/8);
 	printl("");
 
-	if not cmd: # command input was empty 
-		return True; # do nothing
+	# command input was empty 
+	if not cmd:
+		return True; # ignore
 
-	if cmd=="about":
+	#display about help
+	if cmd=="about" or cmd=="a":
 		about();
 		return True;
 
-	if cmd=="?" or cmd=="help":
+	#get command help
+	if cmd=="?" or cmd=="help" or cmd=="h":
 		help(); # prints help text
 		return True;
 
-	if cmd=="post":
+	#post tweet
+	if cmd=="post" or cmd=="p":
 		printl("Type message, maximum 140 characters, newline to finish");
 		printl("Feel free to include you name so others know who's posting");
 		inputs=str(raw_input("> "));
-		if len(inputs) <= 10:
+		if len(inputs) <= 10: #python_twitter wont accept it if its less than 10
 			printl("Message must be more than 10 characters");
-
+									
 			while(1):
 				i=str(raw_input("post anyways ? [Yes/No] ")).lower();
 				if i=="yes":
@@ -178,35 +187,41 @@ def command():
 		if len(inputs) > 140:
 			printl("Message must be less than 140 characters");
 			return True;
+
 		post(inputs);
 		return True;
 
 	#deletes the last post
-	if cmd=="delete":
+	if cmd=="delete" or cmd=="d":
 		while(1):
 			i=str(raw_input("Delete you last post ? [Yes/No]")).lower();
 			if i=="yes":
-				delete(last_status);
+				if last_status!=None:
+					delete(last_status);
+				else:
+					printl("You must post before you can delete your post !");
 				break;
 			if i=="no":
 				break;
 			printl("Please type Yes or No");
 		return True
 
+
 	#undelete's the last post if it was deleted
-	if cmd=="undelete":
+	if cmd=="undelete" or cmd=="u":
 		undelete();
 		return True;
 		
 	#read command
-	if cmd=="read" or cmd=="view":
+	if cmd=="read" or cmd=="view" or cmd=="r":
 		read_replies();
 		return True;
 
 
 	#exit command
-	if cmd=="exit" or cmd=="quit":
+	if cmd=="exit" or cmd=="quit" or cmd=="q":
 		q=str(raw_input("Really [Yes/No]\a")).lower();
+
 		if q=="y" or q=="yes":
 			printl("Goodbye");
 			return False;
@@ -214,7 +229,7 @@ def command():
 			printl("Thank You");
 			return True;
 	
-	bad_cmd();
+	bad_command(); #it wasn't a recognized command
 	return True;
 
 #------------------------------------------------------------
@@ -235,9 +250,9 @@ def post(message):
 		printl("Try explicitly specifying the encoding with the --encoding flag");
 		sys.exit(2);
 
-	limit=api.GetRateLimitStatus();
 	printl(last.user.name+" just posted: "+last.text);
-	#printl(str(limit.get('remaining_hits'))+" posts until limit is reached"); # I'm unsure if this really works ....
+	#limit=api.GetRateLimitStatus(); # I'm unsure if this really works ....
+	#printl(str(limit.get('remaining_hits'))+" posts until limit is reached");
 	#printl("This Resets At: "+str(limit.get('reset_time')));
 	#printl("Hourly Limit: "+str(limit.get('hourly_limit')));
 	time.sleep(status_update_delay/8);
@@ -247,12 +262,15 @@ def post(message):
 #------------------------------------------------------------
 #destroy the last post
 def delete(status):
-	global last_deleted_message;
-	last_deleted_message=status.text;
-	api.DestroyStatus(status.id);
-	printl("Processing ...");
-	time.sleep(status_update_delay/4);
-	printl("Last Post Deleted");
+	if status!=None:
+		global last_deleted_message;
+		last_deleted_message=status.text;
+		api.DestroyStatus(status.id);
+		printl("Processing ...");
+		time.sleep(status_update_delay/4);
+		printl("Last Post Deleted");
+	else:
+		printl("error in delete: status is "+str(status));
 
 def undelete():
 	global last_deleted_message;
@@ -271,21 +289,22 @@ def replies(): #displays # of replies
 def read_replies(): #displays latest replies
 	global last_reply_id;
 
-	replyList=api.GetReplies(None,last_reply_id); #download replies every view of a reply
+	reply_list=api.GetReplies(None,last_reply_id); #download replies every view of a reply
 	
-	replyList.reverse()
+	reply_list.reverse();
 	
-	for reply in replyList:
+	for reply in reply_list:
 
-		printl("    "+reply.user.screen_name+" said: "+reply.text)
+		printl("    "+reply.user.screen_name+" said: "+reply.text);
 		
 		if reply.in_reply_to_status_id != None:
-			printl("In reply to "+reply.in_reply_to_screen_name+":")
-			printl(api.GetStatus(reply.in_reply_to_status_id).text+"\n")
+			printl("In reply to "+reply.in_reply_to_screen_name+":"+api.GetStatus(reply.in_reply_to_status_id).text+"\n");
+		else:
+			printl("In reply to no one ?\n");
 		
-		last_reply_id=reply.id
+		last_reply_id=reply.id;
 		
-		time.sleep(2)
+		time.sleep(2);
 
 #------------------------------------------------------------
 #its the main one ...
@@ -315,15 +334,15 @@ def main(argv):
 
 	'''Main Interactive Command Prompt'''  
 	printl("\a");
-	loopForever = True
-	firstStart = True
+	loop_forever = True;
+	first_start = True;
 
-	while(loopForever):
-		if firstStart:
-			banner()
-			firstStart = False
-		replies()
-		loopForever = command()
+	while(loop_forever):
+		if first_start:
+			banner();
+			first_start = False;
+		replies();
+		loop_forever = command();
 
 
 if __name__ == "__main__":
